TST: component_tests.py
